MOD-1: pillar registry moves server-safe — /modules lives; branded error boundary added
